+++ ivy_tests/test_ivy/helpers/decorators/backend/function_decorator_backend.py
@@ -1,6 +1,5 @@
 from hypothesis import strategies as st
 from typing import Callable, Any
-from ivy_tests.test_ivy.helpers.globals import TestData
 from ivy_tests.test_ivy.helpers.decorators.function_decorator_base import (
     FunctionHandler,
 )
@@ -61,15 +60,6 @@ class BackendFunctionHandler(FunctionHandler):
             "test_data": self.test_data,
             "test_flags": self.test_flags,
         }
-
-    def _build_test_data(self):
-        module_tree, fn_name = self._partition_fn_tree(self.fn_tree)
-        supported_device_dtypes = self._get_supported_devices_dtypes(self.fn_tree)
-        self.test_data = TestData(
-            module_tree=module_tree,
-            fn_name=fn_name,
-            supported_device_dtypes=supported_device_dtypes,
-        )
 
     def _add_test_attributes_to_test_function(self, fn: Callable[..., Any]):
         fn._is_ivy_backend_test = True
