+++ ivy_tests/test_ivy/helpers/decorators/frontend/function_decorator_frontend.py
@@ -3,7 +3,6 @@ from ivy_tests.test_ivy.helpers.decorators.function_decorator_base import (
     FunctionHandler,
 )
 
-from ivy_tests.test_ivy.helpers.structs import FunctionData
 from typing import List, Callable, Any
 from ivy_tests.test_ivy.helpers.test_parameter_flags import (
     build_frontend_function_flags,
@@ -46,15 +45,6 @@ class FrontendFunctionHandler(FunctionHandler):
             generate_frontend_arrays=generate_frontend_arrays,
         )
         self._build_test_data()
-
-    def _build_test_data(self):
-        module_tree, fn_name = self._partition_fn_tree(self.fn_tree)
-        supported_device_dtypes = self._get_supported_devices_dtypes(self.fn_tree)
-        self.test_data = FunctionData(
-            module_tree=module_tree,
-            fn_name=fn_name,
-            supported_device_dtypes=supported_device_dtypes,
-        )
 
     def _build_fn_tree_strategy(self):
         if self.aliases is None:
